split OAuth auth tests into server_auth_oauth_test.go
Mirrors the existing JWE split: extracts all OAuth-related test
functions and helpers (mintSelfIssuedToken, testOAuthProvider,
TestOAuth*, TestValidateAuth, TestResolveOAuthJWKSURL, etc.) from
server_test.go into a dedicated server_auth_oauth_test.go file.
Removes now-unused imports (crypto/rand, crypto/rsa, sync, jose)
from server_test.go.
